Cleaned up Attempt

+++ Installer.sh
@@ -1,15 +1,17 @@
 #!/bin/zsh
 echo -e '\e[90mSearching for the Vogship...'
-curl -s -L --trace-ascii install-dump -o vogship-check https://raw.githubusercontent.com/IsCoffeeTho/42vogship/master/version-check
+curl -L -o vogship-check https://raw.githubusercontent.com/IsCoffeeTho/42vogship/master/version-check
 if grep -q 'vogShip' 'vogship-check';
 then
-  echo -e '\e[97mDownloading the Vogship...'
-  curl -s -L --trace-ascii install-dump -o vogship https://raw.githubusercontent.com/IsCoffeeTho/42vogship/master/commands.sh 1&2>> ./install-dump
-  if grep -q '# vogShip ' "vogship";
+  curl -L -o vogship https://raw.githubusercontent.com/IsCoffeeTho/42vogship/master/commands.sh 1&2>> ./install-dump
+  if grep -q "# vogShip " "vogship";
   then
     echo -e '\e[90m---'
-    echo -e '\e[97mCleaning Environment...'
-    rm -rf ~/.zshrc
+    if grep -q "# vogShip " "~/.zshrc";
+    then
+      echo -e '\e[97mCleaning Environment...'
+      rm -rf ~/.zshrc
+    fi
     echo -e '\e[97mPlacing...'
     mv -f vogship ~/.zshrc
     echo -e '\e[97mVerifying...'
@@ -22,7 +24,13 @@ then
     errorPrint
   fi
 else
-  errorPrint
+  wget -q --spider http://google.com
+  if [ $? -eq 0 ]; then
+      echo -e '\e[91mThere is a problem trying to find vogShip'
+      errorprint
+  else
+      echo -e '\e[91mCannot connect to the Internet'
+  fi
 fi
 rm -rf vogship-check
 rm -rf install-dump
@@ -30,8 +38,6 @@ errorPrint() {
   mv vogShip install-dump
   rm vogship
   cp install-dump download-err.log
-  echo -e '\e[90m---'
-  echo -e '\e[97mThere was a problem downloading the Vogship.'
   echo -e '\e[90m---'
   echo -e '\e[96mIf problems persist, post an issue on the github repo.'
   echo -e '\e[96mJust remember to post your \e[91mdownload-err.log\e[96m file'
